avoid duplicate code
