+++ src/js/app.js
@@ -2,12 +2,6 @@ import { applyTheme } from "./darkmode"
 import { createCopyButton } from "./copycode.js"
 
 import Clipboard from "clipboard"
-
-export const DARK_MODE = "dark"
-export const LIGHT_MODE = "light"
-export const AUTO_MODE = "auto"
-export const THEME = "hugo-geekblog"
-export const TOGGLE_MODES = [AUTO_MODE, DARK_MODE, LIGHT_MODE]
 ;(() => {
   applyTheme()
 })()
