+++ src/js/config.js
@@ -1,0 +1,5 @@
+export const DARK_MODE = "dark"
+export const LIGHT_MODE = "light"
+export const AUTO_MODE = "auto"
+export const THEME = "hugo-geekblog"
+export const TOGGLE_MODES = [AUTO_MODE, DARK_MODE, LIGHT_MODE]
